packages/backend/pyproject.toml
[project]
name = "topogis-linderos-api"
version = "0.1.0"
description = "TopoGIS Linderos - API Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.6",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pyproj>=3.6",
    "python-docx>=1.0",
    "reportlab>=4.0",
    "stripe>=7.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.backends._legacy:_Backend"
